Task 16 third correction

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,15 +9,15 @@
  */
 int main(void)
 {
-	unsigned long long a, b, c, i;
+	long double a, b, c, i;
 
 	a = 1;
 	b = 2;
-	printf("%llu, %llu", a, b);
+	printf("%.0Lf, %.0Lf", a, b);
 	for (i = 2; i < 98; i++)
 	{
 		c = a + b;
-		printf(", %llu", c);
+		printf(", %.0Lf", c);
 		a = b;
 		b = c;
 	}
